src/main/webapp/app/entities/systemuser/update/systemuser-update.component.spec.ts
jest.mock('@angular/router');

import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { of, Subject } from 'rxjs';

import { SystemuserService } from '../service/systemuser.service';
import { ISystemuser, Systemuser } from '../systemuser.model';

import { IUser } from 'app/entities/user/user.model';
import { UserService } from 'app/entities/user/user.service';

import { SystemuserUpdateComponent } from './systemuser-update.component';

describe('Systemuser Management Update Component', () => {
  let comp: SystemuserUpdateComponent;
  let fixture: ComponentFixture<SystemuserUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let systemuserService: SystemuserService;
  let userService: UserService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [SystemuserUpdateComponent],
      providers: [FormBuilder, ActivatedRoute],
    })
      .overrideTemplate(SystemuserUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(SystemuserUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    systemuserService = TestBed.inject(SystemuserService);
    userService = TestBed.inject(UserService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should call User query and add missing value', () => {
      const systemuser: ISystemuser = { id: 456 };
      const user: IUser = { id: 54566 };
      systemuser.user = user;

      const userCollection: IUser[] = [{ id: 6789 }];
      jest.spyOn(userService, 'query').mockReturnValue(of(new HttpResponse({ body: userCollection })));
      const additionalUsers = [user];
      const expectedCollection: IUser[] = [...additionalUsers, ...userCollection];
      jest.spyOn(userService, 'addUserToCollectionIfMissing').mockReturnValue(expectedCollection);

      activatedRoute.data = of({ systemuser });
      comp.ngOnInit();

      expect(userService.query).toHaveBeenCalled();
      expect(userService.addUserToCollectionIfMissing).toHaveBeenCalledWith(userCollection, ...additionalUsers);
      expect(comp.usersSharedCollection).toEqual(expectedCollection);
    });

    it('Should update editForm', () => {
      const systemuser: ISystemuser = { id: 456 };
      const user: IUser = { id: 15817 };
      systemuser.user = user;

      activatedRoute.data = of({ systemuser });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(systemuser));
      expect(comp.usersSharedCollection).toContain(user);
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Systemuser>>();
      const systemuser = { id: 123 };
      jest.spyOn(systemuserService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ systemuser });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: systemuser }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(systemuserService.update).toHaveBeenCalledWith(systemuser);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Systemuser>>();
      const systemuser = new Systemuser();
      jest.spyOn(systemuserService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ systemuser });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: systemuser }));
      saveSubject.complete();

      // THEN
      expect(systemuserService.create).toHaveBeenCalledWith(systemuser);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Systemuser>>();
      const systemuser = { id: 123 };
      jest.spyOn(systemuserService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ systemuser });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(systemuserService.update).toHaveBeenCalledWith(systemuser);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });

  describe('Tracking relationships identifiers', () => {
    describe('trackUserById', () => {
      it('Should return tracked User primary key', () => {
        const entity = { id: 123 };
        const trackResult = comp.trackUserById(0, entity);
        expect(trackResult).toEqual(entity.id);
      });
    });
  });
});
